+++ include/ebpf_structs.h
@@ -10,7 +10,6 @@
 #include <stdbool.h>
 #include <stdint.h>
 #endif
-#include "bpf_structs.h"
 #include "ebpf_windows.h"
 
 #define BPF_ENUM_TO_STRING(X) #X
@@ -154,7 +153,137 @@ typedef enum
     BPF_FUNC_get_current_pid_tgid = 19,      ///< \ref bpf_get_current_pid_tgid
 } ebpf_helper_id_t;
 
+// Cross-platform BPF program types.
+enum bpf_prog_type
+{
+    BPF_PROG_TYPE_UNSPEC, ///< Unspecified program type.
+
+    /** @brief Program type for handling incoming packets as early as possible.
+     *
+     * **eBPF program prototype:** \ref xdp_hook_t
+     *
+     * **Attach type(s):** \ref BPF_XDP
+     *
+     * **Helpers available:** all helpers defined in bpf_helpers.h
+     */
+    BPF_PROG_TYPE_XDP,
+
+    /** @brief Program type for handling socket bind() requests.
+     *
+     * **eBPF program prototype:** \ref bind_hook_t
+     *
+     * **Attach type(s):** \ref BPF_ATTACH_TYPE_BIND
+     *
+     * **Helpers available:** all helpers defined in bpf_helpers.h
+     */
+    BPF_PROG_TYPE_BIND, // TODO(#333): replace with cross-platform program type
+
+    /** @brief Program type for handling various socket operations such as connect(), accept() etc.
+     *
+     * **eBPF program prototype:** \ref sock_addr_hook_t
+     *
+     * **Attach type(s):**
+     *  \ref BPF_CGROUP_INET4_CONNECT
+     *  \ref BPF_CGROUP_INET6_CONNECT
+     *  \ref BPF_CGROUP_INET4_RECV_ACCEPT
+     *  \ref BPF_CGROUP_INET6_RECV_ACCEPT
+     *
+     * **Helpers available:** all helpers defined in bpf_helpers.h
+     */
+    BPF_PROG_TYPE_CGROUP_SOCK_ADDR,
+
+    /** @brief Program type for handling various socket event notifications such as connection established etc.
+     *
+     * **eBPF program prototype:** \ref sock_ops_hook_t
+     *
+     * **Attach type(s):**
+     *  \ref BPF_CGROUP_SOCK_OPS
+     *
+     * **Helpers available:** all helpers defined in bpf_helpers.h
+     */
+    BPF_PROG_TYPE_SOCK_OPS,
+
+    /** @brief Program type for handling calls from the eBPF sample extension. Used for
+     * testing.
+     *
+     * **eBPF program prototype:** see the eBPF sample extension.
+     *
+     * **Attach type(s):** \ref BPF_ATTACH_TYPE_SAMPLE
+     */
+    BPF_PROG_TYPE_SAMPLE = 999
+};
+
+typedef enum bpf_prog_type bpf_prog_type_t;
+
 #define XDP_FLAGS_REPLACE 0x01
+
+// The link type is used to tell which union member is present
+// in the bpf_link_info struct.  There is exactly one non-zero value
+// per union member.
+enum bpf_link_type
+{
+    BPF_LINK_TYPE_UNSPEC, ///< Unspecified link type.
+    BPF_LINK_TYPE_PLAIN,  ///< Normal link type.
+};
+
+enum bpf_attach_type
+{
+    BPF_ATTACH_TYPE_UNSPEC, ///< Unspecified attach type.
+
+    /** @brief Attach type for handling incoming packets as early as possible.
+     *
+     * **Program type:** \ref BPF_PROG_TYPE_XDP
+     */
+    BPF_XDP,
+
+    /** @brief Attach type for handling socket bind() requests.
+     *
+     * **Program type:** \ref BPF_PROG_TYPE_BIND
+     */
+    BPF_ATTACH_TYPE_BIND,
+
+    /** @brief Attach type for handling IPv4 TCP connect() or UDP send
+     * to a unique remote address/port tuple.
+     *
+     * **Program type:** \ref BPF_PROG_TYPE_CGROUP_SOCK_ADDR
+     */
+    BPF_CGROUP_INET4_CONNECT,
+
+    /** @brief Attach type for handling IPv6 TCP connect() or UDP send
+     * to a unique remote address/port tuple.
+     *
+     * **Program type:** \ref BPF_PROG_TYPE_CGROUP_SOCK_ADDR
+     */
+    BPF_CGROUP_INET6_CONNECT,
+
+    /** @brief Attach type for handling IPv4 TCP accept() or on receiving
+     * the first unicast UDP packet from a unique remote address/port tuple.
+     *
+     * **Program type:** \ref BPF_PROG_TYPE_CGROUP_SOCK_ADDR
+     */
+    BPF_CGROUP_INET4_RECV_ACCEPT,
+
+    /** @brief Attach type for handling IPv6 TCP accept() or on receiving
+     * the first unicast UDP packet from a unique remote address/port tuple.
+     *
+     * **Program type:** \ref BPF_PROG_TYPE_CGROUP_SOCK_ADDR
+     */
+    BPF_CGROUP_INET6_RECV_ACCEPT,
+
+    /** @brief Attach type for handling various socket event notifications.
+     *
+     * **Program type:** \ref BPF_PROG_TYPE_SOCK_OPS
+     */
+    BPF_CGROUP_SOCK_OPS,
+
+    /** @brief Attach type implemented by eBPF Sample Extension driver, used for testing.
+     *
+     * **Program type:** \ref BPF_PROG_TYPE_SAMPLE
+     */
+    BPF_ATTACH_TYPE_SAMPLE,
+
+    __MAX_BPF_ATTACH_TYPE,
+};
 
 // Libbpf itself requires the following structs to be defined, but doesn't
 // care what fields they have.  Applications such as bpftool on the other
